requirements.txt
Django~=3.0.5
PyGuitarPro~=0.6
attrs~=19.3.0
django-multiselectfield~=0.1.12